+++ src/main/java/httpserver/http/request/ReadResponseException.java
@@ -1,4 +1,4 @@
-package httpserver.http;
+package httpserver.http.request;
 
 public class ReadResponseException extends RuntimeException {
