+++ src/main/java/httpserver/http/request/RequestSplitter.java
@@ -1,4 +1,4 @@
-package httpserver.http;
+package httpserver.http.request;
 
 public class RequestSplitter {
     public RequestSplitter(String request) {
@@ -7,12 +7,12 @@ public class RequestSplitter {
 
     private final String request;
 
-    String getRequestMethod() {
+    public String getRequestMethod() {
         var splitRequest = splitRequestLine();
         return splitRequest[0];
     }
 
-    String getRequestPath() {
+    public String getRequestPath() {
         var splitRequest = splitRequestLine();
         return splitRequest[1];
     }
