+++ src/test/java/httpserver/http/request/RequestHandlerTest.java
@@ -1,4 +1,4 @@
-package httpserver.http;
+package httpserver.http.request;
 
 import org.junit.jupiter.api.Test;
