+++ src/test/java/httpserver/http/response/ResponseTest.java
@@ -1,5 +1,7 @@
-package httpserver.http;
+package httpserver.http.response;
 
+import httpserver.http.ProtocolVersion;
+import httpserver.http.StatusCode;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
